refactor:#253 온보딩 페이지 루트 경로로 추가 - 기존 메인페이지 /home 경로로 이동 후 path 상수 변경
